RvSearch/requirements.txt
asgiref==3.2.10
Django==3.1
django-appconf==1.0.4
django-compressor==2.4
django-libsass==0.8
django-sass-processor==0.8.1
djangorestframework==3.11.1
libsass==0.20.1
Pillow==7.2.0
pytz==2020.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.3.1
